FooterのMVプレイヤーiframeセットアップの責務分離
